Moved preambleIsAllowed()
