+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -6,7 +6,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
-import static seedu.address.logic.parser.CliSyntax.preambleIsAllowed;
+import static seedu.address.model.person.Type.preambleIsAllowed;
 
 import java.util.Set;
 import java.util.stream.Stream;
