+++ src/main/java/seedu/address/model/person/Type.java
@@ -2,6 +2,7 @@ package seedu.address.model.person;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
+import static seedu.address.logic.parser.CliSyntax.ALLOWED_PREAMBLES;
 
 /**
  * Represents a Person's Type in the address book.
@@ -34,7 +35,21 @@ public class Type {
      * Returns true if a given string is a valid type.
      */
     public static boolean isValidType(String test) {
-        return test.matches(VALIDATION_REGEX);
+        return test.matches(VALIDATION_REGEX) && preambleIsAllowed(test);
+    }
+
+    /**
+     * Checks if a given command uses a preamble that is allowed. (we define preamble as TYPE)
+     * @param preamble
+     * @return true if the preamble is allowed and no if it is not.
+     */
+    public static boolean preambleIsAllowed(String preamble) {
+        for (String s : ALLOWED_PREAMBLES) {
+            if (s.equals(preamble)) {
+                return true;
+            }
+        }
+        return false;
     }
 
     @Override
